+++ packages/lazily/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+	$schema: "http://json.schemastore.org/eslintrc",
+	root: true,
+	extends: ["custom"],
+};