+++ packages/lazily/src/index.ts
@@ -1,0 +1,11 @@
+export const lazily = <Output>(create: () => Output): (() => Output) => {
+	let value: Output | undefined;
+
+	return function factory(): Output {
+		if (value === undefined) {
+			value = create();
+		}
+
+		return value;
+	};
+};