+++ packages/package-template/src/index.ts
@@ -1,0 +1,5 @@
+/**
+ * This is a template for a package.
+ */
+
+export const template = true;