+++ packages/resize-observer-store/src/index.ts
@@ -1,1 +1,10 @@
-export * from "./resize-observer-store";
+import { ResizeObserverStore } from "./resize-observer-store";
+import { lazily } from "@ianduvall/lazily";
+
+/**
+ * Returns the singleton instance of the ResizeObserverStore.
+ * @returns ResizeObserverStore
+ */
+export const getResizeObserverStore = lazily<ResizeObserverStore>(
+	() => new ResizeObserverStore(),
+);