+++ packages/resize-observer-store/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ianduvall/resize-observer-store
 
+## 0.0.5
+
+### Patch Changes
+
+- 9577504: Lazily create store singleton
+
 ## 0.0.4
 
 ### Patch Changes